Addressed sagar's comments
