--- server/src/main/java/org/opensearch/indices/CheckDataResult.java
@@ -1,64 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- */
-
-/*
- * Licensed to Elasticsearch under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-/*
- * Modifications Copyright OpenSearch Contributors. See
- * GitHub history for details.
- */
-
-package org.opensearch.indices;
-
-/**
- * A class used by a CacheTierPolicy to return a result for some data input.
- * The results can be chained together in a short-circuiting way.
- */
-public class CheckDataResult {
-    private boolean isAccepted;
-    private String deniedReason; // null if the data was accepted, has an explanation if data was rejected
-
-    public CheckDataResult(boolean isAccepted, String deniedReason) {
-        this.isAccepted = isAccepted;
-        this.deniedReason = deniedReason;
-    }
-
-    public CheckDataResult composeWith(CheckDataResult other) {
-        if (!this.isAccepted) {
-            return this;
-        } else {
-            return other;
-        }
-    }
-
-    public boolean isAccepted() {
-        return isAccepted;
-    }
-
-    public String getDeniedReason() {
-        return deniedReason;
-    }
-
-}
